requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.12.1
bootstrap-py==2.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.2.0
django-create-react-app==0.9.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
gitdb==4.0.10
GitPython==3.1.32
httpie==3.2.2
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
pguard==0.4.0
psycopg2-binary==2.9.6
Pygments==2.15.1
PyJWT==1.7.1
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
setuptools-scm==7.1.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==7.2.2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.8
sqlparse==0.4.4
sudo==1.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
